[#41] Separate ui model
