+++ android/src/main/java/co/nimblehq/ic/kmm/suv/android/ui/screens/surveydetail/SurveyDetailContentUiModel.kt
@@ -1,0 +1,7 @@
+package co.nimblehq.ic.kmm.suv.android.ui.screens.surveydetail
+
+data class SurveyDetailContentUiModel(
+    val title: String,
+    val description: String,
+    val imageUrl: String
+)
